requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
